lift State in App.js
